+++ Changelog.md
@@ -3,6 +3,8 @@
  - Fix 'password' typo in password-reset controller.
  - Add success message after a user has registered.
  - Add musli and jam-bread recipe.
+ - Add mealtime- and difficulty controller.
+ - Add route to filter recipes in recipe-controller.
 
 ## [1.1.0] - 2020-02-06
 ### Added